Add a generic board game solver and use it to solve BrainVitae
A board game progresses from the given initial state to a final state through a
series of moves. A set of possible moves is applied on a given state of board to
produce a set of resultant states. The generic board game solver, enumerates all
the board states which can result by applying a series of moves to the given
initial state. It does this by finding all the moves that are possible on the
given state, producing the resultant states and then repeating the same process
with all the resultant states. The process is repeated till no new states are
found. The states and the moves are recorded in a database thus being able to
solve board games which have huge number of board states. The solver uses
multiple threads to process multiple states in parallel.

Limitations
1. Right now the board game solver can not handle a board game which has at
least two states reachable from each other. IOW it can not handle board games
characterized by a directed cyclic state graph.
2. For even simple games like tic-tac-toe, which does not have a large number of
states, it uses database. Instead it should use a in-memory store like Map.

Ashutosh Bapat.