Add class representing BrainVitaeBoard
The BrainVitaeBoard object (usually only a single instance in one run of
program) contains two things 1. the information of which cells are used
and which are not used, 2. the bit index of the used cells when
conerting the board state into a bitstring and eventually into an
integer identified.

BVBoardState no more holds information about which cells are used. It
just holds information about whether theAdd class representing
BrainVitaeBoard

The BrainVitaeBoard object (usually only a single instance in one run of
program) contains two things 1. the information of which cells are used
and which are not used, 2. the bit index of the used cells when
conerting the board state into a bitstring and eventually into an
integer identified.

BVBoardState no more holds information about which cells are used. It
just holds information about whether the used cells are filled or empty.
The unused state's state is undefined. used cells are filled or empty.
The unused state's state is undefined.